+++ src/full/Agda/Interaction/AgdaTop.hs
@@ -12,6 +12,7 @@ import Data.Maybe
 import System.IO
 
 import Agda.Interaction.Base
+import Agda.Interaction.ExitCode
 import Agda.Interaction.Response as R
 import Agda.Interaction.InteractionTop
 import Agda.Interaction.Options
@@ -52,8 +53,14 @@ repl callback prompt setup = do
       r <- maybeAbort runInteraction
       case r of
         Done      -> return True -- Done.
-        Error s   -> liftIO (putStrLn s) >> return False
         Command _ -> return False
+        Error s   -> do
+          exit <- optExitOnError <$> commandLineOptions
+          if exit
+            then liftIO (exitAgdaWith CommandError)
+            else do
+              liftIO (putStrLn s)
+              return False
 
     lift Bench.print
     unless done interact'